cv-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.26.0
scikit-image==0.21.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1 